deprecated initial files and implemented ichiro style text field
