+++ src/screens/Feed/index.tsx
@@ -1,0 +1,17 @@
+import styles from "./styles.module.css";
+import IchiroTextField from "./components/IchiroTextField";
+
+
+export default function Feed() {
+
+	return (
+
+		<div class={styles.screen}>
+
+			<IchiroTextField label="Cutoff frequency (Hz)" />
+
+		</div>
+
+	);
+
+}
