+++ src/screens/Feed/styles.module.css
@@ -1,0 +1,12 @@
+
+
+.screen {
+
+	width: 100vw;
+	height: 100vh;
+	background-color: var(--granite);
+	display: flex;
+	justify-content: center;
+	align-items: center;
+
+}
